allow resetDevice to flush the device's pending command stream
The USBL, at least when in TCP mode, will send all pending command/data
on connection. This means that when the driver starts, it receives all
data that was left from previous sessions.

This commit adds the ability for resetDevice to ignore all incoming data
until it gets the acknowledgment from the reset command itself. To
achieve this, we leverage the fact that the USBL device responds to a
command with a string starting with +++ and the command itself.

The commit also adds the validation, in "normal" operations, that the
responses we are seeing actually starts with this expected prefix.